pyproject.toml
[tool.poetry]
name = "tsdm-bot"
version = "0.1.0"
description = "tsdm-bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.15"
nonebot2 = "^2.0.0-beta.1"
beautifulsoup4 = "^4.11.1"
requests-toolbelt = "^0.10.1"
nonebot-adapter-onebot = "^2.1.5"
privatebinapi = "^1.0.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
